requirements.txt
python-telegram-bot==21.6
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
APScheduler==3.10.4
